pyproject.toml
[tool.poetry]
name = "what-to-eat-bot"
version = "0.1.0"
description = ""
authors = ["duxevy <79513394+duxevy@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.15.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
